Add WiFi module with function to reset adapter
This is a module I created about a year ago in response to recurring
problems with hibernating and resuming a laptop that was not fully
supported by Windows 10. The workaround for a non-functional
WiFi adapter was to disable and re-enable it using Device Manager.

This module exposes a function which helps automate that change.
Ironically, newer versions of Windows 10 (Version 1703 in particular)
do not seem to have the same problem as often.

I do not recall how much of the code was inspired or borrowed
from the sources listed in the modules, so see those notes
for background information.